+++ cpp/src/game/Co2dz_Game_Over.h
@@ -28,7 +28,7 @@ public:
 
     // update and draw
     virtual void update (void);
-    virtual void draw (SDL_Renderer & renderer, float lag);
+    virtual void draw (SDL_Renderer & renderer, Uint32 lag);
 };
 
 #endif  //  !defined _CO2DZ_GAME_OVER_H_
